component TodoList separated in two differents components : Todos and TodoList.jsx + CSS added write in white + alert if field todo empty
