+++ front/src/components/TodoList/Todos.css
@@ -1,0 +1,20 @@
+.btnAdd {
+  width:60px;
+  height:30px;
+  text-align:center;
+  color:white;
+  text-decoration:none;
+  background-color: #0155A6;
+  border: none;
+  border-radius: 10%;
+  }
+ 
+.todolist {
+  display: flex;
+  justify-content: center;
+}
+
+.placeholder {
+  margin-right: 40px;
+  width: 500px;
+}
